+++ pallets/subtensor/src/utils/misc.rs
@@ -378,10 +378,10 @@ impl<T: Config> Pallet<T> {
     // ========================
 
     // Per-block epoch cap (dynamic tempo throttle)
-    pub fn get_max_epochs_per_block() -> u32 {
+    pub fn get_max_epochs_per_block() -> u8 {
         MaxEpochsPerBlock::<T>::get()
     }
-    pub fn set_max_epochs_per_block(max_epochs_per_block: u32) {
+    pub fn set_max_epochs_per_block(max_epochs_per_block: u8) {
         MaxEpochsPerBlock::<T>::put(max_epochs_per_block);
         Self::deposit_event(Event::MaxEpochsPerBlockSet(max_epochs_per_block));
     }
